Allow individual bundles to be specified as being in a given region.
If you have a bundle that is not part of a feature it is now possible to
configure the apiregions runtime to make that bundle part of a specified
region. By default bundles that are not in a region can only see the
'global' region.
To make the bundle part of a different region, specify the following
framework property:
  sling.feature.apiregions.bundles=
    {bsn:version}={groupid:artifactId:version};{region},
    {bsn:version}={groupid:artifactId:version};{region}
So the Bundle-SymbolicName and version needs to be mapped to the
associated artifact ID for the bundle artifact and the region needs to
specified.
For example:
  org.foo.bar:1.2.3=org.foo:bar:1.2.3;myregion

The bundle will be added to a synthesized region.