SLING-9004 API Regions runtime should have a configured set of default regions
Make it possible to configure the API Regions runtime with a default set
of regions that all bundles are in, regardless of what feature they are
in or if they are in no feature at all.

To configure the default set of regions set the following framework
property

    sling.feature.apiregions.default=region1,region2

The value of this property is a comma-separated list of regions that all
bundles will be in.

This commit also reverts what was committed for SLING-8976 to use
bundles.properties and features.properties and not
bundleOrigins.properties and regionOrigins.properties.