SLING-9395 Support bundle updates in api regions
The algorithm to find what feature(s) a bundle belongs to at runtime is
now the following:
1. Find the features a bundle belongs to by looking up the
bundle-location to features association.
2. If the bundle location is not associated with any features yet, look
up the bsn+version from the initial, defined list, and associate the
bundle location with the features found there.
This means that if the bundle is later updated, the bundle will still be
associated with the features it was initially associated with and will
therefore still be in the region that it was originally in, even though
its version might have changed. The update must be done through
Bundle.update() or Bundle.update(InputStream) so that the bundle's
location doesn't change..